+++ server/src/main/java/com/krillsson/sysapi/periodictasks/Task.kt
@@ -8,6 +8,7 @@ interface Task {
         RecordDiskReadWriteRate,
         RecordNetworkUploadDownloadRate,
         StoreMetricHistoryEntry,
+        StoreContainerStatisticsHistoryEntry,
         CheckUpdate,
         CheckMonitors
     }